+++ talos-agent/src/bridge.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
-use rclrs::CreateBasicExecutor;
-use talos_common::config::AgentConfig;
+use rclrs::{CreateBasicExecutor, IntoPrimitiveOptions};
+use talos_common::config::{AgentConfig, QosProfile};
 use talos_common::protocol::messages::Response;
 use talos_common::protocol::types::Timestamp;
 use tokio::sync::mpsc;
@@ -9,6 +9,13 @@ use tracing::{info, warn};
 
 use crate::conversions::*;
 use crate::{GraphHandle, JointPublisher};
+
+fn subscription_options<'a>(topic: &'a str, qos: &QosProfile) -> rclrs::PrimitiveOptions<'a> {
+    match qos {
+        QosProfile::Default => topic.into_primitive_options(),
+        QosProfile::SensorData => topic.sensor_data_qos(),
+    }
+}
 
 /// Sender half for ROS 2 callbacks to forward topic data to the router task
 /// without blocking on the router lock.
@@ -27,6 +34,7 @@ pub async fn run(
     for sub_config in &config.subscriptions {
         let topic = sub_config.topic.clone();
         let type_name = sub_config.msg_type.clone();
+        let opts = subscription_options(&topic, &sub_config.qos);
         let tx = bridge_tx.clone();
 
         match sub_config.msg_type.as_str() {
@@ -34,7 +42,7 @@ pub async fn run(
                 let topic_clone = topic.clone();
                 let type_clone = type_name.clone();
                 node.create_subscription::<nav_msgs::msg::Odometry, _>(
-                    &topic,
+                    opts,
                     move |msg: nav_msgs::msg::Odometry| {
                         let stamp = timestamp_from_builtin(&msg.header.stamp);
                         let data = odometry_to_dynvalue(&msg);
@@ -53,7 +61,7 @@ pub async fn run(
                 let topic_clone = topic.clone();
                 let type_clone = type_name.clone();
                 node.create_subscription::<geometry_msgs::msg::Twist, _>(
-                    &topic,
+                    opts,
                     move |msg: geometry_msgs::msg::Twist| {
                         let stamp = Timestamp { sec: 0, nanosec: 0 };
                         let data = twist_msg_to_dynvalue(&msg);
@@ -72,7 +80,7 @@ pub async fn run(
                 let topic_clone = topic.clone();
                 let type_clone = type_name.clone();
                 node.create_subscription::<std_msgs::msg::String, _>(
-                    &topic,
+                    opts,
                     move |msg: std_msgs::msg::String| {
                         let stamp = Timestamp { sec: 0, nanosec: 0 };
                         let data = string_to_dynvalue(&msg);
@@ -91,10 +99,67 @@ pub async fn run(
                 let topic_clone = topic.clone();
                 let type_clone = type_name.clone();
                 node.create_subscription::<sensor_msgs::msg::JointState, _>(
-                    &topic,
+                    opts,
                     move |msg: sensor_msgs::msg::JointState| {
                         let stamp = timestamp_from_builtin(&msg.header.stamp);
                         let data = joint_state_to_dynvalue(&msg);
+                        let response = Response::TopicData {
+                            topic: topic_clone.clone(),
+                            type_name: type_clone.clone(),
+                            stamp,
+                            data,
+                        };
+                        let _ = tx.send(response);
+                    },
+                )?;
+                info!(topic = %topic, msg_type = %type_name, "subscribed");
+            }
+            "sensor_msgs/msg/LaserScan" => {
+                let topic_clone = topic.clone();
+                let type_clone = type_name.clone();
+                node.create_subscription::<sensor_msgs::msg::LaserScan, _>(
+                    opts,
+                    move |msg: sensor_msgs::msg::LaserScan| {
+                        let stamp = timestamp_from_builtin(&msg.header.stamp);
+                        let data = laser_scan_to_dynvalue(&msg);
+                        let response = Response::TopicData {
+                            topic: topic_clone.clone(),
+                            type_name: type_clone.clone(),
+                            stamp,
+                            data,
+                        };
+                        let _ = tx.send(response);
+                    },
+                )?;
+                info!(topic = %topic, msg_type = %type_name, "subscribed");
+            }
+            "sensor_msgs/msg/Imu" => {
+                let topic_clone = topic.clone();
+                let type_clone = type_name.clone();
+                node.create_subscription::<sensor_msgs::msg::Imu, _>(
+                    opts,
+                    move |msg: sensor_msgs::msg::Imu| {
+                        let stamp = timestamp_from_builtin(&msg.header.stamp);
+                        let data = imu_to_dynvalue(&msg);
+                        let response = Response::TopicData {
+                            topic: topic_clone.clone(),
+                            type_name: type_clone.clone(),
+                            stamp,
+                            data,
+                        };
+                        let _ = tx.send(response);
+                    },
+                )?;
+                info!(topic = %topic, msg_type = %type_name, "subscribed");
+            }
+            "geometry_msgs/msg/PoseStamped" => {
+                let topic_clone = topic.clone();
+                let type_clone = type_name.clone();
+                node.create_subscription::<geometry_msgs::msg::PoseStamped, _>(
+                    opts,
+                    move |msg: geometry_msgs::msg::PoseStamped| {
+                        let stamp = timestamp_from_builtin(&msg.header.stamp);
+                        let data = pose_stamped_to_dynvalue(&msg);
                         let response = Response::TopicData {
                             topic: topic_clone.clone(),
                             type_name: type_clone.clone(),
@@ -110,7 +175,7 @@ pub async fn run(
                 let topic_clone = topic.clone();
                 let type_clone = type_name.clone();
                 node.create_subscription::<rcl_interfaces::msg::Log, _>(
-                    &topic,
+                    opts,
                     move |msg: rcl_interfaces::msg::Log| {
                         let stamp = timestamp_from_builtin(&msg.stamp);
                         let data = log_to_dynvalue(&msg);